bootloader: detect/mitigate DVI leakage current halting CODEC init (issue: #81) (PR: #88)
Bug details from https://github.com/apfaudio/tiliqua/issues/81:

```
After some probing, turns out this is a hardware artifact (!!) related to the soft-mute circuit. When a DVI sink (i.e a display) has its 3.3V terminations enabled (usually only a few seconds after an active display connection), it may back-power the Tiliqua through the 50R (sink) + 220R (tiliqua) series resistors into the ECP5. According to the ECP5 datasheet, this is permissible (even if the ECP5 is off). So it should not damage the hardware at least.

The consequence is we see ~1V on the 3.3V rail in this specific scenario, even if the Tiliqua is off. This means the soft-mute circuit never fully deasserts the PDN pin to the audio CODEC when the Tiliqua is powered off, which is required to initialize the CODEC correctly.
```

As a mitigation, in the bootloader, if we detect the CODEC did not initialize correctly, we pulse a zero on PDN and then re-initialize it. This is ONLY performed if we detect the CODEC did not initialize correctly, otherwise it would break soft-mute between bitstreams (where the bootloader also executes).

* bootloader: add AK4619 health detection, hard reset plumbing (does nothing yet)
* bootloader/eurorack_pmod: add flip-flop sequencing
* bootloader: check codec before manifest search
* docs/hw: update hardware connection docs