top/beamrace: add some example beamracing video cores (#142)
Add a simple framework for writing beamracing video cores, where the audio inputs are already correctly synchronized into the pixel clock, making it easy to use the audio inputs to tweak the beamracing pattern.

Most of these are directly translated from https://vga-playground.com/ - with some minor edits so that the audio inputs also do something to the patterns.